# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SchemeSourceType = typeof SchemeSourceType[keyof typeof SchemeSourceType];


export const SchemeSourceType = {
  demo: 'demo',
  verified: 'verified',
} as const;

export interface Scheme {
  id: string;
  name: string;
  nameTe: string;
  nameHi: string;
  summary: string;
  summaryTe: string;
  summaryHi: string;
  description: string;
  descriptionTe: string;
  descriptionHi: string;
  category: string;
  benefits: string[];
  eligibilitySummary: string;
  documents: string[];
  applicationSteps: string[];
  /** @nullable */
  officialUrl: string | null;
  sourceType: SchemeSourceType;
  lastUpdated: string;
  keywords: string[];
  active: boolean;
}

export type SchemeInputSourceType = typeof SchemeInputSourceType[keyof typeof SchemeInputSourceType];


export const SchemeInputSourceType = {
  demo: 'demo',
  verified: 'verified',
} as const;

export interface SchemeInput {
  /**
     * @minLength 1
     * @maxLength 180
     */
  name: string;
  nameTe: string;
  nameHi: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  summary: string;
  summaryTe: string;
  summaryHi: string;
  description: string;
  descriptionTe: string;
  descriptionHi: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  category: string;
  benefits: string[];
  eligibilitySummary: string;
  documents: string[];
  applicationSteps: string[];
  /** @nullable */
  officialUrl?: string | null;
  sourceType: SchemeInputSourceType;
  lastUpdated?: string;
  keywords: string[];
  active?: boolean;
}

export interface Category {
  id: string;
  label: string;
  icon: string;
  count: number;
}

export type EligibilityInputLanguage = typeof EligibilityInputLanguage[keyof typeof EligibilityInputLanguage];


export const EligibilityInputLanguage = {
  en: 'en',
  te: 'te',
  hi: 'hi',
} as const;

export interface EligibilityInput {
  /**
     * @minimum 0
     * @maximum 120
     */
  age?: number;
  gender?: string;
  state?: string;
  rural?: boolean;
  occupation?: string;
  /** @minimum 0 */
  annualIncome?: number;
  farmer?: boolean;
  student?: boolean;
  disability?: boolean;
  seniorCitizen?: boolean;
  housingNeed?: boolean;
  employmentStatus?: string;
  language?: EligibilityInputLanguage;
}

export interface EligibilityMatch {
  scheme: Scheme;
  relevance: number;
  reasons: string[];
}

export type AssistantInputLanguage = typeof AssistantInputLanguage[keyof typeof AssistantInputLanguage];


export const AssistantInputLanguage = {
  en: 'en',
  te: 'te',
  hi: 'hi',
} as const;

export interface AssistantInput {
  /**
     * @minLength 1
     * @maxLength 1000
     */
  message: string;
  language: AssistantInputLanguage;
}

export interface AssistantResponse {
  answer: string;
  schemes: Scheme[];
}

export interface ReportInput {
  schemeId: string;
  /**
     * @minLength 3
     * @maxLength 1000
     */
  message: string;
}

export interface ReportResponse {
  id: number;
  received: boolean;
}

export interface AdminSummary {
  total: number;
  active: number;
  categories: number;
}

export interface AdminLoginInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
  demoMode: boolean;
}

export type ListSchemesParams = {
q?: string;
category?: string;
sort?: ListSchemesSort;
};

export type ListSchemesSort = typeof ListSchemesSort[keyof typeof ListSchemesSort];


export const ListSchemesSort = {
  relevance: 'relevance',
  name: 'name',
} as const;

export type SearchSchemesParams = {
/**
 * @minLength 1
 */
q: string;
};