# lib/api-zod/src/generated/types/assistantInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantInputLanguage } from './assistantInputLanguage';

export interface AssistantInput {
  /**
     * @minLength 1
     * @maxLength 1000
     */
  message: string;
  language: AssistantInputLanguage;
}