# lib/api-zod/src/generated/types/assistantResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Scheme } from './scheme';

export interface AssistantResponse {
  answer: string;
  schemes: Scheme[];
}