# lib/api-zod/src/generated/types/eligibilityMatch.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Scheme } from './scheme';

export interface EligibilityMatch {
  scheme: Scheme;
  relevance: number;
  reasons: string[];
}