# lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReportInput {
  schemeId: string;
  /**
     * @minLength 3
     * @maxLength 1000
     */
  message: string;
}